Separate device and uncaged code to new packages
