+++ go.mod
@@ -1,6 +1,7 @@
 module github.com/shermp/Kobo-UNCaGED
 
 require (
+	github.com/BurntSushi/toml v0.3.1
 	github.com/bamiaux/rez v0.0.0-20170731184118-29f4463c688b
 	github.com/gofrs/uuid v3.2.0+incompatible
 	github.com/kapmahc/epub v0.1.1
