+++ kobo-uncaged/device/types_test.go
@@ -1,4 +1,4 @@
-package main
+package device
 
 import (
 	"path/filepath"
